# openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "PatentFlow BE API",
    "description" : "PatentFlow FE 우선 연동 API 문서",
    "version" : "v1"
  },
  "servers" : [
    {
      "url" : "http://localhost",
      "description" : "Generated server url"
    }
  ],
  "paths" : {
    "/api/v1/settings/valuation-criteria" : {
      "get" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "getValuationCriteria",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseValuationCriteriaResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "updateValuationCriteria",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ValuationCriteriaRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseValuationCriteriaResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/valuation-criteria/prompts/{axis}" : {
      "get" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "getValuationCriteriaPrompt",
        "parameters" : [
          {
            "name" : "axis",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseValuationPromptResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "updateValuationCriteriaPrompt",
        "parameters" : [
          {
            "name" : "axis",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ValuationPromptUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseValuationPromptResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/review-quarters/{quarterKey}" : {
      "put" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "updateQuarterSetting",
        "parameters" : [
          {
            "name" : "quarterKey",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/QuarterSettingRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseQuarterSettingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/review-periods/{periodNumber}" : {
      "put" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "updatePeriodTemplate",
        "parameters" : [
          {
            "name" : "periodNumber",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "integer",
              "format" : "int32"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReviewPeriodTemplateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseReviewPeriodTemplateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/fee-rules/{country}" : {
      "put" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "updateFeeRule",
        "parameters" : [
          {
            "name" : "country",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/FeeRuleUpdateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseFeeRuleResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/country-extensions/{country}" : {
      "put" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "updateCountryExtension",
        "parameters" : [
          {
            "name" : "country",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CountryExtensionRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseCountryExtensionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/classifications/{type}/{value}" : {
      "put" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "renameClassification",
        "parameters" : [
          {
            "name" : "type",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "value",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClassificationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseClassificationResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "deleteClassification",
        "parameters" : [
          {
            "name" : "type",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "name" : "value",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseClassificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/business-checklist-items/{itemId}" : {
      "put" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "updateBusinessChecklistItem",
        "parameters" : [
          {
            "name" : "itemId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BusinessChecklistItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBusinessChecklistItemResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "deleteBusinessChecklistItem",
        "parameters" : [
          {
            "name" : "itemId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/review-quarters/{quarterKey}/activate" : {
      "post" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "activateQuarter",
        "parameters" : [
          {
            "name" : "quarterKey",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseQuarterActivateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/classifications/{type}" : {
      "post" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "addClassification",
        "parameters" : [
          {
            "name" : "type",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ClassificationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseClassificationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/business-checklist-items" : {
      "get" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "getBusinessChecklistItems",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListBusinessChecklistItemResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "createBusinessChecklistItem",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BusinessChecklistItemRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseBusinessChecklistItemResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/response-deadline" : {
      "get" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "getResponseDeadline",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseResponseDeadlineResponse"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "updateResponseDeadline",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResponseDeadlineResponse"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseResponseDeadlineResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/mail-lead-months" : {
      "get" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "getMailLeadMonths",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMailLeadMonthsResponse"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "updateMailLeadMonths",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MailLeadMonthsResponse"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseMailLeadMonthsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/settings/review-quarters" : {
      "get" : {
        "tags" : [
          "settings-controller"
        ],
        "operationId" : "getQuarterSettings",
        "parameters" : [
          {
            "name" : "year",
            "in" : "query",
            "required" : false,
            "schema" : {
              "type" : "integer",
              "format" : "int32",
              "default" : 0
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListQuarterSettingResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ValuationCriteriaRequest" : {
        "type" : "object",
        "properties" : {
          "axisWeights" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "number",
              "format" : "double"
            }
          },
          "gradeCutoffs" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "number",
              "format" : "double"
            }
          },
          "maintainThreshold" : {
            "type" : "number",
            "format" : "double"
          },
          "subscoreWeights" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object",
              "additionalProperties" : {
                "type" : "integer",
                "format" : "int32"
              }
            }
          }
        },
        "required" : [
          "axisWeights",
          "gradeCutoffs",
          "maintainThreshold",
          "subscoreWeights"
        ]
      },
      "ApiResponseValuationCriteriaResponse" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ValuationCriteriaResponse"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ValuationPromptUpdateRequest" : {
        "type" : "object",
        "properties" : {
          "markdown" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "expectedChecksum" : {
            "type" : "string"
          }
        }
      },
      "ApiResponseValuationPromptResponse" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/ValuationPromptResponse"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "QuarterSettingRequest" : {
        "type" : "object",
        "properties" : {
          "startDate" : {
            "type" : "string",
            "format" : "date"
          },
          "endDate" : {
            "type" : "string",
            "format" : "date"
          },
          "submissionDeadline" : {
            "type" : "string",
            "format" : "date"
          },
          "businessResponseDueDate" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "ApiResponseQuarterSettingResponse" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/QuarterSettingResponse"
          },
          "message" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "ReviewPeriodTemplateRequest" : {
        "type" : "object",
        "properties" : {
          "startMonth" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 12,
            "minimum" : 1
          },
          "startDay" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 31,
            "minimum" : 1
          },
          "endMonth" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 12,
            "minimum" : 1
          },
          "endDay" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 31,
            "minimum" : 1
          }
        },
        "required" : [
          "endDay",
          "endMonth",
          "startDay",
          "startMonth"
        ]
      },
...
      "CountryExtensionRequest" : {
        "type" : "object",
        "properties" : {
          "extensionMonths" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 240,
            "minimum" : 0
          }
        }
      },
...
    }
  }
}